pyproject.toml
[tool.poetry]
name = "r-a-m"
version = "0.1.0"
description = ""
authors = ["zifrit <zifrit@list.ru>"]
readme = "README.md"
packages = [{include = "r"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
aiosqlite = "^0.20.0"
alembic = "^1.13.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
